+++ src/components/FaceRecognition/styles.js
