+++ src/components/Rank/styles.js
