+++ src/pages/MainPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const MainPage = () => {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default MainPage
