--- src/styles/Logo.styles.js
@@ -1,6 +1,0 @@
-import styled from 'styled-components'
-
-export const Img = styled.img`
-  overflow: auto;
-  width: 75px;
-`
